backend/requirements.txt
# GitFlow-CMS Backend Dependencies
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# GitHub Integration
PyGithub==2.2.0

# Code Analysis & Formatting
black==24.2.0

# AI Integration
groq==0.4.2

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.1

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0
